% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_charts.R
\name{calculate_hexbins_from_shots}
\alias{calculate_hexbins_from_shots}
\title{Converting shot location data into hexbins for plotting}
\usage{
calculate_hexbins_from_shots(
  shots,
  league_averages = NULL,
  binwidths = c(1, 1),
  min_radius_factor = 0.6,
  fg_diff_limits = c(-0.15, 0.15),
  fg_pct_limits = c(0.2, 0.7),
  fg_freq_limits = c(-0.075, 0.075),
  pps_limits = c(0.5, 1.5)
)
}
\arguments{
\item{shots}{shot location data from play-by-play source}

\item{league_averages}{league average accuracy from shot location areas}

\item{binwidths, min_radius_factor}{hex chart sizing specifications}

\item{fg_diff_limits, fg_freq_limits, fg_pct_limits, pps_limits}{values to bound
shot chart legend}
}
\description{
Converting shot location data into hexbins for plotting
}
\examples{
\dontrun{
calculate_hexbins_from_shots(shots = shots,
league_averages = league_averages, ...)
}

}
