% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_charts.R
\name{convert_to_avgs}
\alias{convert_to_avgs}
\title{Convert shots to averages to apply to 'league_averages' input for hex chart}
\usage{
convert_to_avgs(shots)
}
\arguments{
\item{shots}{shot location data from play-by-play source}
}
\description{
Convert shots to averages to apply to 'league_averages' input for hex chart
}
\examples{
\dontrun{
convert_to_avgs(shots = shots)
}

}
