% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_colors.R
\name{find_cbb_team_colors}
\alias{find_cbb_team_colors}
\title{Find the two primary colors for a college basketball team or program}
\usage{
find_cbb_team_colors(team, switch_col = F)
}
\arguments{
\item{team}{name of the team for which to return their logo}

\item{switch_col}{logical of whether or not to switch the order of
the returned colors}
}
\description{
Find the two primary colors for a college basketball team or program
}
\examples{
\dontrun{
find_cbb_team_colors(team = "Duke", switch_col = F)
}

}
