% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_colors.R
\name{find_cbb_team_logo}
\alias{find_cbb_team_logo}
\title{Find the logo url for a college basketball team or program}
\usage{
find_cbb_team_logo(team)
}
\arguments{
\item{team}{name of the team for which to return their logo}
}
\description{
Find the logo url for a college basketball team or program
}
\examples{
\dontrun{
find_cbb_team_logo(team = "Duke")
}

}
