% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_charts.R
\name{generate_hex_chart}
\alias{generate_hex_chart}
\title{Generate a hex chart given data produced from \code{calculate_hexbins_from_shots}}
\usage{
generate_hex_chart(
  hex_shots,
  type = sym(c("bounded_fg_diff", "bounded_freq_diff", "bounded_fg_pct",
    "bounded_points_per_shot")),
  team = "Duke",
  opponent = F,
  low_alpha_range = 0.85
)
}
\arguments{
\item{hex_shots}{shot location data produced from \code{calculate_hexbins_from_shots}}

\item{type}{type of shot chart to create, one of sym(c("bounded_fg_diff",
"bounded_freq_diff", "bounded_fg_pct", "bounded_points_per_shot"))}

\item{team}{team for which the shot location data is from}

\item{opponent}{logical denoting if shots come from the team or their opponent's}

\item{low_alpha_range}{value denoting the lower limit of alpha for ggplot points}
}
\description{
Generate a hex chart given data produced from \code{calculate_hexbins_from_shots}
}
\examples{
\dontrun{
generate_hex_chart(hex_shots = duke, team = "Duke", opponent = F,
type = sym("bounded_fg_diff"), low_alpha_range = 0.85)
}

}
