% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacklich10-package.R
\docType{package}
\name{jacklich10-package}
\alias{jacklich10}
\alias{jacklich10-package}
\title{jacklich10: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
