% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_colors.R
\name{join_cbb_team_info}
\alias{join_cbb_team_info}
\title{Join in team information for a college basketball team or program}
\usage{
join_cbb_team_info(data, join_name)
}
\arguments{
\item{data}{dataframe with college basketball data and team names}

\item{join_name}{column in which to join team information on}
}
\description{
Join in team information for a college basketball team or program
}
\examples{
\dontrun{
dplyr::tibble(team = "Duke") \%>\%
join_cbb_team_info(data = ., join_name = "team")
}

}
