% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_charts.R
\name{team_player_hex_chart}
\alias{team_player_hex_chart}
\title{Generate a hex chart for a given team or player}
\usage{
team_player_hex_chart(
  shots,
  league_averages,
  team,
  player = NULL,
  type = dplyr::sym(c("bounded_fg_diff", "bounded_freq_diff", "bounded_fg_pct",
    "bounded_points_per_shot")),
  ...
)
}
\arguments{
\item{shots}{shot location data from play-by-play source}

\item{league_averages}{league average accuracy from shot location areas}

\item{team}{team for which the shot location data is from}

\item{player}{player for which the shot location data is from (default is NULL)}

\item{type}{type of shot chart to create, one of sym(c("bounded_fg_diff",
"bounded_freq_diff", "bounded_fg_pct", "bounded_points_per_shot"))}

\item{...}{other arguments passed to \code{calculate_hexbins_from_shots}}
}
\description{
Generate a hex chart for a given team or player
}
\examples{
\dontrun{
team_player_hex_chart(shots = duke, league_averages = league_averages,
team = "Duke", player = "Matthew Hurt", type = sym("bounded_fg_diff"),
binwidths = c(2, 2), min_radius_factor = 0.05)
}

}
