% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_mov_avg.R
\name{wt_mov_avg}
\alias{wt_mov_avg}
\title{Weighted rolling averages with a dynamic window}
\usage{
wt_mov_avg(var, weight, window, type, moving = T)
}
\arguments{
\item{var}{vector of data to take rolling average}

\item{weight}{vector of data to weight rolling average with}

\item{window}{length of rolling average window}

\item{type}{one of \code{type = c("s", "t", "w", "m", "e", "r")} from \code{pracma::movavg}}

\item{moving}{logical specifying if rolling average should have a dynamic window or not}
}
\description{
Weighted rolling averages with a dynamic window
}
\examples{
\dontrun{

wt_mov_avg(var = fgm, weight = fga, window = window, type = "s", moving = T)

}

}
