% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacklich-package.R
\docType{package}
\name{jacklich-package}
\alias{jacklich}
\alias{jacklich-package}
\title{jacklich: Personal R Package}
\description{
Loading in my commonly used functions and themes.
}
\author{
\strong{Maintainer}: Jack Lichtenstein \email{jacklich10@gmail.com}

}
\keyword{internal}
